Move combination form types into sub namespace
